+++ 52_24_elimina_bonificacions/Lloguer.java
@@ -31,11 +31,11 @@ public class Lloguer {
         }
         return quantitat;
     }
-    public double bonificacionsDeLloguer(Lloguer lloguer) {
+    public int bonificacionsDeLloguer(Lloguer lloguer) {
         if(getVehicle().getCategoria() == Vehicle.Categoria.LUXE && getDies()>1) {
-            return 2.0;
+            return 2;
         }
-        return 1.0;
+        return 1;
     }
 
 }
